requirements.txt
astroid==2.6.5
attrs==20.3.0
Authlib==0.15.4
autopep8==1.5.7
awscli==1.20.8
backports.entry-points-selectable==1.1.0
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.9.3
blinker==1.4
blis==0.7.4
botocore==1.21.8
braceexpand==0.1.7
cachelib==0.2.0
cairocffi==1.2.0
CairoSVG==2.5.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
check50==3.3.2
click==7.1.2
cloudmersive-nlp-api-client==4.0.1
cloudmersive-validate-api-client==3.2.3
colorama==0.4.3
compare50==1.2.4
cryptography==3.4.7
cs50==7.0.2
cssselect2==0.4.1
cycler==0.10.0
cymem==2.0.5
defusedxml==0.7.1
detectlanguage==1.5.0
distlib==0.3.2
docutils==0.15.2
EditorConfig==0.12.3
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.1.0/en_core_web_md-3.1.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
filelock==3.0.12
Flask==1.1.4
Flask-Mail==0.9.1
Flask-Session==0.4.0
greenlet==1.1.0
help50==3.0.4
html5lib==1.1
icdiff==1.9.1
idna==3.2
ikp3db==1.4.5
Inflector==2.0.12
intervaltree==2.1.0
isort==5.9.3
itsdangerous==1.1.0
jellyfish==0.8.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsbeautifier==1.14.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lib50==2.0.9
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
murmurhash==1.0.5
natsort==7.1.1
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
pathy==0.6.0
pexpect==4.8.0
Pillow==8.3.1
platformdirs==2.1.0
plotly==5.1.0
pprintpp==0.4.0
preshed==3.0.5
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pylint==2.9.6
pylint-django==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
pyspellchecker==0.6.2
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
render50==5.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3cmd==2.1.0
s3transfer==0.5.0
six==1.16.0
smart-open==5.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.1.1
spacy-legacy==3.0.8
SQLAlchemy==1.4.22
sqlparse==0.4.1
srsly==2.4.1
style50==2.7.5
submit50==3.0.4
tenacity==8.0.1
termcolor==1.1.0
textblob==0.15.3
textstat==0.7.1
thinc==8.0.8
tinycss2==1.1.0
toml==0.10.2
tqdm==4.61.2
twython==3.9.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.6.0
wasabi==0.8.2
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
